# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Ruta - API de verificacion de entregas",
    "description" : "Verificacion de entregas con PIN, foto y ubicacion. Todas las rutas /api/v1/admin/** y /api/v1/driver/** exigen la cookie HttpOnly de sesion (\"access_token\") que POST /api/v1/auth/login establece via Set-Cookie. Las peticiones que modifican datos requieren ademas el header X-XSRF-TOKEN (proteccion CSRF con doble cookie). Swagger UI envia la cookie solo si el login se hizo desde el propio navegador con \"Try it out\".",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Mismo origen que la aplicacion"
  } ],
  "security" : [ {
    "sessionCookie" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/driver/invoices" : {
      "get" : {
        "tags" : [ "Entregas del conductor" ],
        "summary" : "Buscar facturas pendientes de entrega",
        "description" : "Facturas publicadas, con PIN habilitado y aun no confirmadas, filtradas por numero o cliente.",
        "operationId" : "search",
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Facturas que coinciden con la busqueda (hasta 20)",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/InvoiceResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/driver/invoices/{id}/lines" : {
      "get" : {
        "tags" : [ "Entregas del conductor" ],
        "summary" : "Lineas de una factura",
        "description" : "Productos vendidos en la factura, para el checklist de entrega del conductor.",
        "operationId" : "lines",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Lineas de la factura",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/InvoiceLineResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "InvoiceResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "number" : {
            "type" : "string"
          },
          "partnerName" : {
            "type" : "string"
          },
          "deliveryAddress" : {
            "type" : "string"
          },
          "invoiceDate" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string",
            "enum" : [ "draft", "posted", "cancel" ]
          },
          "expectedLatitude" : {
            "type" : "number",
            "format" : "double"
          },
          "expectedLongitude" : {
            "type" : "number",
            "format" : "double"
          }
        }
      },
      "InvoiceLineResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "description" : {
            "type" : "string"
          },
          "quantity" : {
            "type" : "number",
            "format" : "double"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "sessionCookie" : {
        "type" : "apiKey",
        "name" : "access_token",
        "in" : "cookie"
      }
    }
  }
}